Supports creation of any swift object
Since I simplified the public interface to take Anything as the
binding object (even if only Classes and Strings are used), I
might as well support the option to return Anything as well.

Note that it still remains the case that if you want to injector a class
that does not inherit from NSObject (or is NSObject) - or if you want to
use a special initializer of any Class, then you must set a custom binding.